refactored code: formatting and syntax corrections were applied.
