--- forecastTemplate.json
@@ -1,55 +1,0 @@
-{
-    "city": "Armenia",
-    "country": "Colombia",
-    "tempCelsius": 30,
-    "condition": "Partly cloudy",
-    "conditionIcon": " //cdn.weatherapi.com/weather/64x64/day/116.png",    
-    "tempFeels": 33.3,
-    "uv": 5,
-    "forecastToday": [
-        {
-        "tempCelsius": 15.7,
-        "condition": "Clear",
-        "conditionIcon": "//cdn.weatherapi.com/weather/64x64/night/113.png",
-        "uv": 1
-        },
-        {
-        "tempCelsius": 14.7,
-        "condition": "Patchy rain possible",
-        "conditionIcon": " //cdn.weatherapi.com/weather/64x64/day/176.png",  
-        "uv": 3
-        },
-        {
-        "tempCelsius": 15.7,
-        "condition": "Clear",
-        "conditionIcon": "//cdn.weatherapi.com/weather/64x64/night/113.png",
-        "uv": 1
-        },
-        {
-        "tempCelsius": 14.7,
-        "condition": "Patchy rain possible",
-        "conditionIcon": " //cdn.weatherapi.com/weather/64x64/day/176.png",  
-        "uv": 3
-        }
-    ],
-    "forecast": [
-        { 
-            "date": "2023-04-25", 
-            "weather": {
-                "tempCelsius": 15.7,
-                "condition": "Clear",
-                "conditionIcon": "//cdn.weatherapi.com/weather/64x64/night/113.png",
-                "uv": 1
-            } 
-        },
-        { 
-            "date": "2023-04-26", 
-            "weather": {
-                "tempCelsius": 14.7,
-                "condition": "Patchy rain possible",
-                "conditionIcon": " //cdn.weatherapi.com/weather/64x64/day/176.png",  
-                "uv": 3 
-            }
-        }
-    ]
-}
