+++ html/forecast.html
@@ -1,18 +1,18 @@
 <html>
   <head>
     <title>Forecast weather</title>
-    <link rel="stylesheet" type="text/css" href="./forecast.css" />
+    <link rel="stylesheet" type="text/css" href="../style/forecast.css" />
     <style>
       @import url('https://fonts.googleapis.com/css2?family=Lato&family=Open+Sans:wght@300&display=swap');
     </style>
     <style>
       @import url('https://fonts.googleapis.com/css2?family=Roboto:wght@300&display=swap');
     </style>
-    <script type="module" src="./forecast.js"></script>
+    <script type="module" src="../scripts/forecast.js"></script>
   </head>
   <body>
     <header>
-      <a href="./index.html">MAIN</a>
+      <a href="../index.html">MAIN</a>
       <a href="./history.html">HISTORY</a>
     </header>
 
@@ -49,7 +49,6 @@
               </div>
             </div>
           </div>
-        
         </div>
       </div>
