+++ html/history.html
@@ -2,13 +2,13 @@
 <html>
     <head>
         <title>Historical weather</title>
-        <link rel="stylesheet" type="text/css" href="./history.css" />
+        <link rel="stylesheet" type="text/css" href="../style/history.css" />
         <script type="text/javascript" src="https://www.gstatic.com/charts/loader.js"></script>
-        <script type="module" src="./history.js"></script>
+        <script type="module" src="../scripts/history.js"></script>
     </head>
     <body>
         <header>
-            <a href="./index.html">MAIN</a>
+            <a href="../index.html">MAIN</a>
             <a href="./forecast.html">FORECAST</a>
         </header>
