+++ index.html
@@ -13,8 +13,8 @@
                 <h1>WEATHERPOINT</h1>
 
                 <div class="buttonsContainer">
-                    <a href="./forecast.html">FORECAST</a>
-                    <a href="./history.html">HISTORY</a>
+                    <a href="./html/forecast.html">FORECAST</a>
+                    <a href="./html/history.html">HISTORY</a>
                 </div>
             </div>
         </section>
